[FAB-10611] Adding nexus release thirdparty jobs
This patchset add two new jobs for releasing
the thirdparty images to the nexus repo.

+++ jjb/fabric-sdk-node/shell/include-raw-publish-doc.sh
@@ -1,22 +1,25 @@
-#!/bin/bash -eu
+#!/bin/bash -e
 set -o pipefail
 
-cd ${WORKSPACE}/gopath/src/github.com/hyperledger/fabric-sdk-node
-# Generate sdk docs
-gulp doc
-SDK_COMMIT=$(git rev-parse --short HEAD)
-TARGET_REPO=$NODE_SDK_USERNAME.github.io.git
-git clone https://github.com/$NODE_SDK_USERNAME/$TARGET_REPO
-
-# Remove API docs to Target repository
-rm -rf $NODE_SDK_USERNAME.github.io/*
-
-# Copy API docs to Target repository
-cp -r docs/gen/* $NODE_SDK_USERNAME.github.io
-cd $NODE_SDK_USERNAME.github.io
-git add .
-git commit -m "SDK commit - $SDK_COMMIT"
-git config remote.gh-pages.url https://$NODE_SDK_USERNAME:$NODE_SDK_PASSWORD@github.com/$NODE_SDK_USERNAME/$TARGET_REPO
-
-# Push API docs to Target repository
-git push gh-pages master
+ARCH=$(uname -m)
+echo "--------> ARCH:" $ARCH
+if [ "$ARCH" != "s390x" ]; then
+    cd ${WORKSPACE}/gopath/src/github.com/hyperledger/fabric-sdk-node
+    # Generate SDK-Node API docs
+    gulp doc
+    # Short Head commit
+    SDK_COMMIT=$(git rev-parse --short HEAD)
+    TARGET_REPO=$NODE_SDK_USERNAME.github.io.git
+    # Clone SDK_NODE API doc repository
+    git clone https://github.com/$NODE_SDK_USERNAME/$TARGET_REPO
+    # Remove API docs target repository
+    rm -rf $NODE_SDK_USERNAME.github.io/*
+    # Copy API docs to target repository & push to gh-pages URL
+    cp -r docs/gen/* $NODE_SDK_USERNAME.github.io
+    cd $NODE_SDK_USERNAME.github.io
+    git add .
+    git commit -m "SDK commit - $SDK_COMMIT"
+    git config remote.gh-pages.url https://$NODE_SDK_USERNAME:$NODE_SDK_PASSWORD@github.com/$NODE_SDK_USERNAME/$TARGET_REPO
+    # Push API docs to Target repository
+    git push gh-pages master
+fi
